/recorded でも scroll 位置を保存できるようにした
